LayoutTests/fast/xpath/nsresolver-exception-expected.txt

CONSOLE MESSAGE: line 14: Value undefined (result of expression window.foobar) is not object.
SUCCESS
